Split sample browser row paint widgets
